Cleanup current api views.
