+++ project/api/utils.py
@@ -1,8 +1,26 @@
-from decimal import Decimal
+import json
 import datetime
 import collections
-import json
-from django.http import HttpResponse
+from decimal import Decimal
+from django.http import HttpResponse, Http404
+
+from django.shortcuts import get_object_or_404
+
+from api.models import Account
+
+
+def get_account(user=None, pk=None, username=None):
+    """ gets author based on the user """
+    if user:
+        return get_object_or_404(Account, user=user)
+    elif pk:
+        return get_object_or_404(Account, pk=pk)
+    elif username:
+        return get_object_or_404(Account, user__username=username)
+
+    else:
+        raise Http404
+
 
 def json_custom_parser(obj):
     """
