+++ project/api/views/__init__.py
@@ -1,0 +1,4 @@
+from .account import AccountViewSet
+from .category import CategoryViewSet
+from .civi import CiviViewSet
+from .thread import ThreadViewSet
