+++ project/api/write.py
@@ -16,8 +16,8 @@ from api.forms import UpdateProfileImage
 from api.models import Thread
 from api.tasks import send_mass_email
 from models import Account, Activity, Category, Civi, CiviImage, Invitation
-from utils.custom_decorators import require_post_params
-from utils.constants import US_STATES
+from core.custom_decorators import require_post_params
+from core.constants import US_STATES
 
 @login_required
 @require_post_params(params=['title', 'summary', 'category_id'])
