+++ project/core/string_templates.py
