+++ project/frontend_views/views.py
@@ -10,8 +10,8 @@ from django.template.response import TemplateResponse
 
 from api.models import Category, Account, Thread, Civi, Activity, Invitation
 from api.forms import UpdateProfileImage
-from utils.constants import US_STATES
-from utils.custom_decorators import beta_blocker, login_required, full_account
+from core.constants import US_STATES
+from core.custom_decorators import beta_blocker, login_required, full_account
 
 
 def base_view(request):
